Refactor color utilities
