+++ public/js/index.js
@@ -1,3 +1,5 @@
+import { getRandomColor, isColorDark } from '../utils/color.js';
+
 const dataAttributeLabels = {
     pos:            'part of speech',
     entType:        'type',
@@ -27,13 +29,6 @@ const dataAttributeLabels = {
     likeNum:        'like number',
     likeUrl:        'like url',
 };
-
-function isColorDark(hex) {
-    const r = parseInt(hex.substring(1, 3), 16);
-    const g = parseInt(hex.substring(3, 5), 16);
-    const b = parseInt(hex.substring(5, 7), 16);
-    return 0.299 * r + 0.587 * g + 0.114 * b < 128;    
-}
 
 function adjustMainBody() {
     const headerHeight = document.querySelector('.header').offsetHeight;
