+++ views/index.ejs
@@ -8,8 +8,7 @@
     <title>Upload and Visualize XML</title>
     <link rel="stylesheet" 
           href="/css/styles.css">    
-    <script src="/js/index.js" 
-            defer></script>
+    <script type="module" src="/js/index.js"></script>
 </head>
 
 <body>
